Create README.md
Tor-browser-with-patch-installed I did not create the Tor-Browser.
In contrast, I have added the patch to the #! tor-browser file so you can run it as root in Linux at your own risk!
Warranty Disclaimer: You use the code and suggestions at your own risk! Besides, you waive the right to tort claims and liabilities by downloading
Using the code is at the sole discretion considering you waive the right to liability claims for its use or distribution.